Adds kevin makes googleoAuth its own component
